Got the latest build working on Linux x86 1.  Removed a few compile errors, such as a useless cast. 2.  Links in pthreads where needed 3.  Updated SConstructs so they can compile ARM (not fully working yet)

+++ Notebook/Common/SCSV/Test/SCSVParserTest.cpp
@@ -163,7 +163,7 @@ BOOST_AUTO_TEST_CASE(SCSVParser_blankFileTest){
 
     BOOST_CHECK_EQUAL(status.errorNumber, Core::SCSV_OKAY);
 
-    BOOST_CHECK_EQUAL(status.CSVValues.size(), 0);
+    BOOST_CHECK_EQUAL(status.CSVValues.size(), static_cast<unsigned int>(0));
 }
 
 ///\brief tests the case where there is no blank line at the end of the file
